test projs

+++ ConsolePinnacle/Models/BalanceModel.cs
@@ -4,9 +4,9 @@ namespace ConsolePinnacle.Models
 {
     public class BalanceModel
     {
-        public float availableBalance { get; set; }
-        public float outstandingTransactions { get; set; }
-        public float givenCredit { get; set; }
+        public double availableBalance { get; set; }
+        public double outstandingTransactions { get; set; }
+        public double givenCredit { get; set; }
         public string currency { get; set; }
     }
 }
