+++ SteamApiWeb/ViewModels/Bets/BalanceViewModel.cs
@@ -2,9 +2,9 @@
 {
     public class BalanceViewModel
     {
-        public float AvailableBalance { get; set; }
-        public float OutstandingTransactions { get; set; }
-        public float GivenCredit { get; set; }
+        public double AvailableBalance { get; set; }
+        public double OutstandingTransactions { get; set; }
+        public double GivenCredit { get; set; }
         public string Currency { get; set; }
     }
 }
