+++ SteamApiWeb/ViewModels/Bets/OddsViewModel.cs
@@ -25,33 +25,33 @@ namespace SteamBetterWeb.ViewModels.Bets
             public int LineId { get; set; }
             public int Number { get; set; }
             public DateTime Cutoff { get; set; }
-            public float MaxMoneyline { get; set; }
+            public double MaxMoneyline { get; set; }
             public Moneyline Moneyline { get; set; }
-            public float MaxSpread { get; set; }
-            public float MaxTotal { get; set; }
+            public double MaxSpread { get; set; }
+            public double MaxTotal { get; set; }
             public Spread[] Spreads { get; set; }
             public Total[] Totals { get; set; }
         }
 
         public class Moneyline
         {
-            public float Home { get; set; }
-            public float Away { get; set; }
+            public double Home { get; set; }
+            public double Away { get; set; }
         }
 
         public class Spread
         {
-            public float Hdp { get; set; }
-            public float Home { get; set; }
-            public float Away { get; set; }
+            public double Hdp { get; set; }
+            public double Home { get; set; }
+            public double Away { get; set; }
             public long AltLineId { get; set; }
         }
 
         public class Total
         {
-            public float Points { get; set; }
-            public float Over { get; set; }
-            public float Under { get; set; }
+            public double Points { get; set; }
+            public double Over { get; set; }
+            public double Under { get; set; }
             public long AltLineId { get; set; }
         }
